--HG--
branch : hotfixes-for-v-1-6-2


Former-commit-id: 1748850f1f19a87703b2addb7899dff4985c515e

+++ src/Core/Infrastructure/WB.Core.Infrastructure.Raven/Implementation/ReadSide/RavenReadSideService.cs
@@ -43,8 +43,8 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
             UpdateStatusMessage("No administration operations were performed so far.");
         }
 
-        public RavenReadSideService(IStreamableEventStore eventStore, IEventDispatcher eventBus, DocumentStore ravenStore, ILogger logger, IRavenReadSideRepositoryWriterRegistry writerRegistry,
-        IReadSideRepositoryCleanerRegistry cleanerRegistry)
+        public RavenReadSideService(IStreamableEventStore eventStore, IEventDispatcher eventBus, DocumentStore ravenStore, ILogger logger,
+            IRavenReadSideRepositoryWriterRegistry writerRegistry, IReadSideRepositoryCleanerRegistry cleanerRegistry)
         {
             this.eventStore = eventStore;
             this.eventBus = eventBus;
@@ -75,9 +75,11 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
                 GetReadableErrors());
         }
 
-        public void RebuildAllViewsAsync()
+        public void RebuildAllViewsAsync(int skipEvents)
         {
-            new Task(this.RebuildAllViews).Start();
+            skipEvents = Math.Max(0, skipEvents);
+
+            new Task(() => this.RebuildAllViews(skipEvents)).Start();
         }
 
         public void RebuildViewsAsync(string[] handlerNames)
@@ -120,7 +122,7 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
             }
         }
 
-        private void RebuildAllViews()
+        private void RebuildAllViews(int skipEvents)
         {
             if (!areViewsBeingRebuiltNow)
             {
@@ -128,7 +130,7 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
                 {
                     if (!areViewsBeingRebuiltNow)
                     {
-                        this.RebuildAllViewsImpl();
+                        this.RebuildAllViewsImpl(skipEvents);
                     }
                 }
             }
@@ -201,14 +203,17 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
             return this.writerRegistry.GetAll().Where(w => viewTypes.Contains(w.ViewType)).ToArray();
         }
 
-        private void RebuildAllViewsImpl()
+        private void RebuildAllViewsImpl(int skipEvents)
         {
             try
             {
                 areViewsBeingRebuiltNow = true;
 
-                this.DeleteAllViews();
-                this.CleanUpAllWriters();
+                if (skipEvents == 0)
+                {
+                    this.DeleteAllViews();
+                    this.CleanUpAllWriters();
+                }
 
                 string republishDetails = "<<NO DETAILS>>";
 
@@ -216,7 +221,7 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
                 {
                     this.EnableCacheInAllRepositoryWriters();
 
-                    republishDetails = this.RepublishAllEvents();
+                    republishDetails = this.RepublishAllEvents(skipEvents);
                 }
                 finally
                 {
@@ -382,9 +387,9 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
             UpdateStatusMessage("Cache in repository writers disabled.");
         }
 
-        private string RepublishAllEvents()
+        private string RepublishAllEvents(int skipEventsCount = 0)
         {
-            int processedEventsCount = 0;
+            int processedEventsCount = skipEventsCount;
             int failedEventsCount = 0;
 
             ThrowIfShouldStopViewsRebuilding();
@@ -398,17 +403,17 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
             DateTime republishStarted = DateTime.Now;
             UpdateStatusMessage(
                 "Acquiring first portion of events. "
-                + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount));
+                + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount));
 
-            foreach (CommittedEvent[] eventBulk in this.eventStore.GetAllEvents())
+            foreach (CommittedEvent[] eventBulk in this.eventStore.GetAllEvents(skipEvents: skipEventsCount))
             {
                 foreach (CommittedEvent @event in eventBulk)
                 {
-                    ThrowIfShouldStopViewsRebuilding();
+                    ThrowIfShouldStopViewsRebuilding(GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount));
 
                     UpdateStatusMessage(
                         string.Format("Publishing event {0}. ", processedEventsCount + 1)
-                        + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount));
+                        + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount));
 
                     try
                     {
@@ -432,45 +437,47 @@ namespace WB.Core.Infrastructure.Raven.Implementation.ReadSide
 
                 UpdateStatusMessage(
                     "Acquiring next portion of events. "
-                    + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount));
+                    + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount));
             }
 
             logger.Info(String.Format("Processed {0} events, failed {1}", processedEventsCount, failedEventsCount));
 
             UpdateStatusMessage(string.Format("All events were republished. "
-                + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount)));
+                + GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount)));
 
-            return GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount);
+            return GetReadablePublishingDetails(republishStarted, processedEventsCount, allEventsCount, failedEventsCount, skipEventsCount);
         }
 
         private static string GetReadablePublishingDetails(DateTime republishStarted,
-            int processedEventsCount, int allEventsCount, int failedEventsCount)
+            int processedEventsCount, int allEventsCount, int failedEventsCount, int skippedEventsCount)
         {
+            int republishedEventsCount = processedEventsCount - skippedEventsCount;
+
             TimeSpan republishTimeSpent = DateTime.Now - republishStarted;
 
             int speedInEventsPerMinute = (int)(
                 republishTimeSpent.TotalSeconds == 0
                 ? 0
-                : 60 * processedEventsCount / republishTimeSpent.TotalSeconds);
+                : 60 * republishedEventsCount / republishTimeSpent.TotalSeconds);
 
             TimeSpan estimatedTotalRepublishTime = TimeSpan.FromMilliseconds(
-                processedEventsCount == 0
+                republishedEventsCount == 0
                 ? 0
-                : republishTimeSpent.TotalMilliseconds / processedEventsCount * allEventsCount);
+                : republishTimeSpent.TotalMilliseconds / republishedEventsCount * allEventsCount);
 
             return string.Format(
-                "Processed events: {1}. Total events: {2}. Failed events: {3}.{0}Time spent republishing: {4}. Speed: {5} events per minute. Estimated time: {6}.",
+                "Processed events: {1}. Total events: {2}. Skipped events: {3} Failed events: {4}.{0}Time spent republishing: {5}. Speed: {6} events per minute. Estimated time: {7}.",
                 Environment.NewLine,
-                processedEventsCount, allEventsCount, failedEventsCount,
+                processedEventsCount, allEventsCount, skippedEventsCount, failedEventsCount,
                 republishTimeSpent.ToString(@"hh\:mm\:ss"), speedInEventsPerMinute, estimatedTotalRepublishTime.ToString(@"hh\:mm\:ss"));
         }
 
-        private static void ThrowIfShouldStopViewsRebuilding()
+        private static void ThrowIfShouldStopViewsRebuilding(string readableStatus = "")
         {
             if (shouldStopViewsRebuilding)
             {
                 shouldStopViewsRebuilding = false;
-                throw new Exception("Views rebuilding stopped by request.");
+                throw new Exception("Views rebuilding stopped by request. " + readableStatus);
             }
         }
